refactor: extract option normalization in preset-typescript
