+++ packages/babel-preset-typescript/src/index.js
@@ -1,35 +1,19 @@
 import { declare } from "@babel/helper-plugin-utils";
 import transformTypeScript from "@babel/plugin-transform-typescript";
-import { OptionValidator } from "@babel/helper-validator-option";
-const v = new OptionValidator("@babel/preset-typescript");
+import normalizeOptions from "./normalize-options.js";
 
 export default declare((api, opts) => {
   api.assertVersion(7);
 
   const {
+    allExtensions,
     allowDeclareFields,
     allowNamespaces,
+    isTSX,
     jsxPragma,
+    jsxPragmaFrag,
     onlyRemoveTypeImports,
-  } = opts;
-
-  const jsxPragmaFrag = v.validateStringOption(
-    "jsxPragmaFrag",
-    opts.jsxPragmaFrag,
-    "React.Fragment",
-  );
-
-  const allExtensions = v.validateBooleanOption(
-    "allExtensions",
-    opts.allExtensions,
-    false,
-  );
-
-  const isTSX = v.validateBooleanOption("isTSX", opts.isTSX, false);
-
-  if (isTSX) {
-    v.invariant(allExtensions, "isTSX:true requires allExtensions:true");
-  }
+  } = normalizeOptions(opts);
 
   const pluginOptions = isTSX => ({
     allowDeclareFields,
